Fixed issues/111:Sidecars keep reloading even no endpoints change #111

+++ pkg/sidecar/providers/pipy/client/types.go
@@ -32,7 +32,7 @@ var (
 
 // Codebase is Pipy's logic unit
 type Codebase struct {
-	Version     uint64   `json:"version,omitempty"`
+	Version     uint32   `json:"version,omitempty"`
 	Path        string   `json:"path,omitempty"`
 	Main        string   `json:"main,omitempty"`
 	Base        string   `json:"base,omitempty"`
